Added Species

+++ src/main.rs
@@ -21,7 +21,7 @@ pub use fvec2::*;
 
 fn main() -> GameResult {
     let window_config = load_config::<WindowConfig>("window")?;
-    let window_mode = WindowMode::default().dimensions(window_config.width, window_config.height);
+    let window_mode = WindowMode::default().dimensions(window_config.width as f32, window_config.height as f32);
 
     let (mut ctx, event_loop) = ContextBuilder::new("Physarum", "nicholasVilela")
         .window_mode(window_mode)
